optimization/requirements.txt
ortools>=9.6
psycopg2-binary
python-dotenv
scikit-learn
numpy
google-cloud-pubsub
